requirements.txt
anthropic==0.97.0
google-api-core==2.30.3
google-api-python-client==2.194.0
google-auth==2.49.2
google-auth-httplib2==0.3.1
google-auth-oauthlib==1.3.1
googleapis-common-protos==1.74.0
gspread==6.2.1
python-dotenv==1.2.2
requests==2.33.1
requests-oauthlib==2.0.0
yfinance==1.2.2
